[BugFix] -> Moved quite some variables to "long long" (exp, Exp, AddDamage...) or "unsigned long long", basically all that are involved into exp and monster HP, due to Turak boss, -> for cf monsters, it seems lvl and exp were inverted, -> added a check if a monster had too much hp (for turak boss only lol),
[Bug]
-> boss turak life bar wrong displayed, but don't know how to do it for now,

+++ World Server/WorldMonster.h
@@ -76,7 +76,8 @@ class CMonster : public CCharacter
       	bool PlayerInRange( );
       	bool PlayerInGrid( );      //LMA: maps
       	CPlayer* GetNearPlayer( UINT mdist = 20 );
-        void AddDamage( CCharacter* enemy, long int hitpower);
+        //void AddDamage( CCharacter* enemy, long int hitpower);
+        void AddDamage( CCharacter* enemy, long long hitpower);
         CDrop* GetDrop( );
 
         //
@@ -92,7 +93,10 @@ class CMonster : public CCharacter
         unsigned int GetCritical( );
         unsigned int GetAttackSpeed( );
         unsigned int GetMoveSpeed( );
-        unsigned int GetMaxHP( );
+
+        //unsigned int GetMaxHP( );
+        unsigned long long GetMaxHP( );
+
         float GetAttackDistance( );
 
         // Events
